This PR achieve **x2.88 speedup** for 4-thread, and it's **x3.65 speedup** in compare to current version. benchmark result as follow:

|  Threads  |   Time   | Speedup 1 | Speedup 2 |
|:---------:|:--------:|:---------:|:---------:|
|  current  | 59.6 sec |     -     | baseline  |
| 1 thread  | 47.3 sec | baseline  |   x1.26   |
| 2 threads | 35.9 sec |   x1.32   |   x1.66   |
| 3 threads | 21.2 sec |   x2.23   |   x2.80   |
| 4 threads | 16.4 sec |   x2.88   |   x3.65   |

(1 GB log, 5.6 M lines, M2 Air, average of 5 runs)

## Method

The simple control flow is as follows:
```python
for each file:
  read file
  for each line:
    for every specifier in format ('%h %^[%d:%t %^] "%r" %s ...'):
      parse token into LogItem    <- parallelize here
    process LogItem to Metric     <- uplift this function
```

Since `parse_specifier()` takes much time here and doesn't have data races with others, we parallelize here first.

```python
for each file:
  read file
  while not eof:
    for each line:
      save lines to Job1
    create threads:
      for every line in Job1:
        for every specifier in format ('%h %^[%d:%t %^] "%r" %s ...'):
          parse token into LogItem    <- parallelize here
          save LogItem to Job2
    for every LogItem in Job2:
      process LogItem to Metric     <- uplift this function
```

## Major speedup causes

### Parallelize `parse_log()`

Parsing logs takes about 70% of execution time, we parallelize this part.

And let `process_log()` run along with `parse_log()` to achieve x3.0 speedup for 5-thread machine.

### Use `GLog->start_time` instead of `localtime_r()`

When solving the data race in `init_log_item()`, we introduced `start_time` in `GLog`.
And find out that `parse_specifier()` (which will be executed about 10 * <line count> times, depends on log format) also called this function every time.

Copying `struct tm` instead of convert Unix Timestamp to {year, month, day, hour, ...} saved a lot of time.

This contributed x1.15 speedup for serial version.

### Change `GLogItem->status` from `char*` to `int`

When we measure the execute time of each modules, we find out that "Status Code" module takes 30% among them.

The reason is finding the `{"404" => "404 - Not Found"}` mapping with foreach() and string compare consume too many time. So we changed `status_code` to integer, and use `status_codes[404]` to lookup.

This will speedup process_log() by x1.38, and have x1.28 overall speedup if `parse_log()` in the bottleneck.

## Details

Instead of `--multi-thread-parsing` we talked about before, using `-j, --jobs` (same as `make`) might be more reasonable for users. The default option `conf.jobs=1` will not create new threads for parse/process log.

We split it into 20 commits:
- Prepare
    - Merge `read_lines()` for with/without `GETLINE`
    - Rename `pre_process_log()` to `parse_line()`
    - Uplift `process_log()` from `parse_line()`
    - Uplift `process_log()` from `read_line()`
    - Add --jobs and --chunk-size options
- Parse in batch
    - Replace logitems with GJob in `read_lines()`
    - Replace lines with GJob in `read_lines()`
    - Parse lines to GLogItem in batch
- Solve data races
    - Eliminate data races with `__sync_bool_compare_and_swap()`
    - Use `__sync_add_and_fetch()`
    - Add GLog->start_time to prevent using `localtime_r()` per GLogItem
    - Use logitem instead of glog->item for `init_log_item()`
- Parallelize
    - Implement the parallel version for log parsing
    - Process logs in batch
    - Run proc_logs while read_lines_thread in progress
    - Change GLogItem->status from `char *` to `int`
- Misc
    - Change `char *` to `const char *` for parser
    - Add multi-thread to README
    - Update .gitignore

With every commits:
- exactly the same output as master branch
- have no compiling warning
- passed the Valgrind test
- conform to coding style
- have no big change in compare to previous one, hope it would be easy to review step by step

## Future Work

- Parallelize `process_log()` by separating modules into different threads
    - I didn't think of a simple way to balance the thread count between `parse_logs()` and `process_logs()`
    - And which thread parses which modules, the workload varies for modules
- Decide `CHUNK_SIZE` more scientifically
    - Not considered about low memory machines or highly efficient servers yet
- Thread pool / Job queue
- Check for rarely happened crashes
    - During the development stage, we encountered many crashes too, but this version seems to be fine

Fixes #377