@Sea-n 2023-12-18
Hi Gerardo,

We find that processing status code is time consuming, and achieved x1.2 speedup.
|  Version  | Overall Time | proc_log() | "Status Code" Module |
|:---------:|:--------:|:---------:|:---------:|
|  Current  | 20.6 sec | 17.4 sec  |  6.2 sec  |
| Improved  | 16.2 sec | 12.6 sec  |  1.4 sec  |

And as you advised, we extracted `CHUNK_SIZE` to `--chunk-size` option now, it's easier then we though!

This PR is ready for review now. 😊