@Sea-n 2023-12-19
The code is updated! Here is the summary/reason.

### 1-a. jobs count
It now has `--jobs=<1-6>` and recommended 4-thread in the help text.
The hard limit is set to `--jobs=12`.

### 1-b. chunk-size
It's a long part, leave it to the end.

### 2-a. error message
Oh, I forgot to check if `logitem->errstr` exists and just replaced it.
Now it will display the original error message.

### 2-b. dry run
I've thinking about whether should we run strictly 10 items or one round (4 thread * chunk 1024 = 4096 items).

Compared to the added complexity to the codebase (and maybe slow down a little bit because of the if-else), I think 10 ms waste of time at the startup is acceptable.

In the extreme case, 6 threads * chunk 8,192 ~= 50,000 lines, it should be done in 0.x seconds.

### 3-a. memory leak
Changed from `cnt == num_tests` to `cnt >= num_tests`, and free the `logitem` before `return NULL`.
How can I miss this 😅

### 3-b. memory leak
Related to (2-a) error message.
It was overwritten, so no one freed the original `errstr`.

## Chunk Size

The benchmark:

![512 to 262144](https://github.com/allinurl/goaccess/assets/10102616/f075258c-c54e-4cff-8199-ef5545ba82da)

(Note that the log file only has 5,600,000 lines)

My test command: `for i in {1..5}; do for c in 1024 4096 16384 65536 262144  512 2048 8192 32768 131072; do for j in {1..4}; do time ./goaccess access.log.1gb -o /dev/null --jobs=$j --chunk-size=$c; sleep 1; done; done; done 2>&1 | tee -a benchmark-result`

And for the extreme values:

![1 to 1024](https://github.com/allinurl/goaccess/assets/10102616/9a814b26-cf06-46d5-b854-4fe9a4a22315)

The overhead for low chunk size might be:
- pthread_create() and pthread_join() without thread pool
- function call for every chunk

![16384 to 1048576](https://github.com/allinurl/goaccess/assets/10102616/87087c04-4d27-4144-9282-60d4d608a327)

The overhead for extremely high chunk size:
- The idle time for the first round (have no logitems to process)
- The final process time (with no lines need to be parsed)

So I add a hard limit along with recommended range, and temporarily keep the default value untouched.
It's hard to decide the max/min hard/soft limit and the default value, it's welcome to change to any number you like.

#### For 1 thread
For 1 thread, chunk-size has no significant difference. (Though higher `chunk-size` will have a slightly speedup.)

We have some options here:
- only check for `--jobs=4 --chunk-size=8`, allow `--chunk-size=8 --jobs=4` to work
- also check the `conf.chunk_size` in the `--jobs` option handling
- check `conf.chunk_size` in the main program (after all options are parsed)
- still imply a meanless but harmless limit for `--jobs=1`
And I chose the last one as my first thought, it's also up to you to decide.