@allinurl 2023-12-20
This is fantastic! Thank you for the effort put into this; it's a significant enhancement to the parsing process. Merged.