@allinurl 2023-12-20
whoops, I rushed the merge. Looks like something's broken when piping in data. e.g., 

    tail -f access.log | goaccess - --log-format=VCOMBINED -o report.html